requirements.txt
fastapi
uvicorn
requests
chromadb
onnxruntime
transformers
python-dotenv
redis
numpy